@TimothyMakkison 2023-01-01
> Art the officially unsupported targets (pre .NET 6) worth to vectorized them? As at the moment only .NET 6+ is officially supported I wouldn't bother with .NET Core 3.1 optimizations anymore.
> There's quite some code duplication.

Good point, it probably isn't worth supporting them, I'll remove them unless the owner want to keep them. Should remove some bloat. I'll see if I can create one or two helper functions.

Could I ask you a question? From what I gathered from testing .Net 3, 3.1, 5 should have constants declared outside the method and .Net 7 should have constants delcared inline. Finally .Net 6 uniquely runs the same with constants declared outside or inline. Is this correct?