@gfoidl 2023-01-02
.NET 6 onwards should use the vector constants inline. The JIT has then the ability to optimize the usage. It streamlines the C# code also.
Before .NET 6 it was better to have the vector constants declared outside.
See https://github.com/dotnet/runtime/pull/54827 (and the linked issue) for more info.