@hadashiA 2023-08-18
Sorry for the late reply and thanks for the great PR.
We have confirmed the changes and would be happy to merge them.
We would also like to adopt your suggestion about adding a test environment.
Thanks a lot :)