@hadashiA 2023-08-19
There seems to be some discussion left unanswered.

> Modernise the code to use C# 11 - use file scoped namespaces, simplify slices.
> Enable nullable across the project and update relevant methods.

Currently, we want the same code to work in Unity 2019.3 so we may not change it until we change the supported version.

> Consider adding comparison operators <, <=, >, >=.
> Add 'NotNullWhen' attributes to methods such as 'TryParse'.
> public string ToBase64 uses ArrayPool, stackalloc could instead be used here. The SkipLocalsInit could be used to remove the overhead of clearing the span.

This is probably a reasonable change.
We would like to consider this.

> Would the logic used in the original get Time and internal Ulid(long timestampMilliseconds... work on big endian machines?

Hmm, it looks like you are right. I would like to check with the original author.

> Experimental change to `internal Ulid(long,..)`

Although we do not want to add too much complexity to the value generation for the sake of speed, `BinaryPrimitives.ReverseEndianness` may be reasonable.

Thanks for the suggestions. further pr is welcome.